Core/Game: forgotten changes from r3cc82905662712385f0d03895a426310e6cc96bd
--HG--
branch : trunk

+++ src/server/game/Maps/MapManager.cpp
@@ -269,19 +269,19 @@ void MapManager::Update(uint32 diff)
     for (; iter != i_maps.end(); ++iter)
     {
         if (m_updater.activated())
-            m_updater.schedule_update(*iter->second, i_timer.GetCurrent());
+            m_updater.schedule_update(*iter->second, uint32(i_timer.GetCurrent()));
         else
-            iter->second->Update(i_timer.GetCurrent());
+            iter->second->Update(uint32(i_timer.GetCurrent()));
     }
     if (m_updater.activated())
         m_updater.wait();
 
     for (iter = i_maps.begin(); iter != i_maps.end(); ++iter)
-        iter->second->DelayedUpdate(i_timer.GetCurrent());
+        iter->second->DelayedUpdate(uint32(i_timer.GetCurrent()));
 
-    sObjectAccessor.Update(i_timer.GetCurrent());
+    sObjectAccessor.Update(uint32(i_timer.GetCurrent()));
     for (TransportSet::iterator iter = m_Transports.begin(); iter != m_Transports.end(); ++iter)
-        (*iter)->Update(i_timer.GetCurrent());
+        (*iter)->Update(uint32(i_timer.GetCurrent()));
 
     i_timer.SetCurrent(0);
 }
